# docs/saas-openapi.yaml
openapi: 3.0.3
info:
  title: Champion SaaS Customer API
  description: >
    Server-to-server HTTP bridge between the Champion website and the
    authoritative Go SaaS repositories. See docs/SAAS_API.md for the full
    narrative handoff (auth model, acting-user contract, examples, tenant
    scoping). This file is a machine-readable companion to that document,
    not a replacement for it.
  version: "1.0.0"
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        default: https://dayzkillfeedbot-production.up.railway.app
        description: CHAMPION_SAAS_API_URL
security:
  - serviceAuth: []
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      description: >
        Authorization: Bearer <CHAMPION_SAAS_API_SECRET> - the same value as
        the bot's WEBSITE_API_SECRET. Server-side only; never expose to
        browser code.
  parameters:
    ActingUser:
      name: X-Champion-Acting-User
      in: header
      required: true
      description: >
        The Discord user ID of the person acting on this request. Must have
        been synced at least once via POST /api/saas/users/sync. Not
        required on that one route.
      schema:
        type: string
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        discordUserId: { type: string }
        discordUsername: { type: string }
        discordGlobalName: { type: string }
        avatar: { type: string }
    UserSyncRequest:
      type: object
      required: [discord_user_id, discord_username]
      properties:
        discord_user_id: { type: string }
        discord_username: { type: string }
        discord_global_name: { type: string }
        avatar: { type: string }
...
    VerifyPermissionsRequest:
      type: object
      required: [channelId]
      properties:
        channelId: { type: string }
...
  responses:
    Unauthorized:
      description: Missing/invalid service auth, or acting user not synced.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    InvalidRequest:
      description: Malformed body, invalid ID, or an invalid setup-progress transition.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    RateLimited:
      description: Too many requests for this route/caller.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
paths:
  /api/saas/users/sync:
    post:
      summary: Sync safe Discord identity fields (never an OAuth token).
      security: [{ serviceAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserSyncRequest" }
      responses:
        "200":
          description: Synced user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSummary" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/saas/organizations:
    get:
      summary: List organizations the acting user belongs to.
      parameters: [{ $ref: "#/components/parameters/ActingUser" }]
      responses:
        "200":
          description: Organizations.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrganizationSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create an organization + OWNER membership atomically.
      parameters: [{ $ref: "#/components/parameters/ActingUser" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrganizationRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationSummary" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/saas/organizations/{organizationID}:
    get:
      summary: Get one organization (membership required).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      responses:
        "200":
          description: Organization.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/saas/organizations/{organizationID}/dashboard:
    get:
      summary: One-call customer dashboard summary.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      responses:
        "200":
          description: Dashboard summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/saas/organizations/{organizationID}/installations:
    post:
      summary: Create an installation (OWNER/ADMIN).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInstallationRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationSummary" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/saas/organizations/{organizationID}/installations/{installationID}:
    get:
      summary: Get one installation.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      responses:
        "200":
          description: Installation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/setup:
    get:
      summary: Get resumable setup progress.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      responses:
        "200":
          description: Setup progress.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupProgress" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update setup progress (OWNER/ADMIN). Step order enforced server-side.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupProgressPatchRequest" }
      responses:
        "200":
          description: Updated setup progress.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupProgress" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/saas/organizations/{organizationID}/discord/guilds/eligible:
    post:
      summary: Re-derive eligibility for OAuth-verified guild candidates.
      description: >
        botInstalled comes from the bot's local gateway state cache (zero
        Discord REST calls per candidate) - never a live per-guild lookup.
        The final selected guild is still live-verified by
        /discord/connection and /discord/verify-installation.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EligibleGuildsRequest" }
      responses:
        "200":
          description: Eligibility results.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DiscordGuildSummary" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/saas/organizations/{organizationID}/discord/connection:
    post:
      summary: Persist a verified Discord guild selection (OWNER/ADMIN).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectGuildRequest" }
      responses:
        "200":
          description: Persisted connection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiscordGuildConnectionSummary" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503":
          description: Discord bot session unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/discord/verify-installation:
    post:
      summary: Verify live bot presence in the connected guild; may advance NOT_STARTED to DISCORD_CONNECTED.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      responses:
        "200":
          description: Verification result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiscordInstallationVerification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503":
          description: Discord bot session unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/discord/verify-permissions:
    post:
      summary: Check live bot channel permissions (View Channel, Send Messages, Embed Links, Read Message History).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyPermissionsRequest" }
      responses:
        "200":
          description: Permission verification result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiscordPermissionVerification" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: Installation not yet Discord-verified.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503":
          description: Discord bot session unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/saas/organizations/{organizationID}/nitrado/connect:
    post:
      summary: Validate a Nitrado account token live and store its encrypted credential envelope (OWNER/ADMIN).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NitradoConnectRequest" }
      responses:
        "200":
          description: Connected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NitradoConnectResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/NitradoUnavailable" }
  /api/saas/organizations/{organizationID}/nitrado/services:
    get:
      summary: Discover the connected account's supported DayZ console services (PlayStation + Xbox only).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
      responses:
        "200":
          description: Supported services.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NitradoServiceSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: No Nitrado connection exists yet for this organization.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503": { $ref: "#/components/responses/NitradoUnavailable" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/dayz-server:
    post:
      summary: Select a discovered service as the installation's DayZ server (OWNER/ADMIN). Re-verifies live before persisting.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SelectDayZServerRequest" }
      responses:
        "200":
          description: >
            Selected, or reused (see SelectDayZServerResponse.reusedInstallation)
            when this guild+server pair already belonged to a different
            existing installation in the same organization.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SelectDayZServerResponse" }
        "400":
          description: Invalid request, or the service is not a supported DayZ console platform.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Installation not found, or the service isn't on the connected Nitrado account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: This game_servers row is already owned by a different organization.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503": { $ref: "#/components/responses/NitradoUnavailable" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/dayz-server/validate:
    post:
      summary: Safe, read-only live check that the selected DayZ server is still reachable and supported.
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      responses:
        "200":
          description: Validation result (reachable/supported reflect reality even when false - this is a status check, not an error).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DayZServerValidation" }
        "400":
          description: No DayZ server has been selected for this installation yet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/saas/organizations/{organizationID}/installations/{installationID}/discord/channels:
    get:
      summary: List the installation's guild's Champion-selectable Discord channels (text/announcement, bot-viewable only).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      responses:
        "200":
          description: Channels, sorted by position then name.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DiscordChannelSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: Discord bot session unavailable, or the channel list could not be loaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    post:
      summary: Create a new Discord channel for Customize mode (OWNER/ADMIN).
      parameters:
        - { $ref: "#/components/parameters/ActingUser" }
        - { $ref: "#/components/parameters/organizationID" }
        - { $ref: "#/components/parameters/installationID" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDiscordChannelRequest" }
      responses:
        "201":
          description: Channel created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiscordChannelSummary" }
        "400":
          description: Invalid channel name, or categoryId does not belong to this installation's guild.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: Discord bot session unavailable, missing Manage Channels, or channel creation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...